+++ src/components/FilteredGallery/filtered-gallery.scss
