+++ src/components/ImageDisplay/display.scss
