+++ src/components/SocialMediaLinks/social.scss
@@ -1,0 +1,21 @@
+
+.social {
+  
+  padding-left: 1em;
+
+  a {
+    text-decoration: none;
+    color: #000;
+  }
+  a:hover {
+    color: #036;
+  }
+
+  img, svg {
+    margin-right: 5px;
+  }
+  img {
+    height: 2rem;
+  }
+  
+}
